scripts/kong-api-scripts/requirements.txt
requests
retry
argparse
jwt
logging
cryptography 
PyJWT
PyYAML
